+++ NEWS.md
@@ -1,5 +1,6 @@
 # incase (development version)
 
+* Fixed bug where `*_case()` would return `NULL` if no conditions were `TRUE`.
 * Added `tibble` to suggested packages.
 
 # incase 0.3.0